Crash recovery (#114)
* Subscriber server

* Nits

* Nits

* Add store to consensus

* Fix main & tests
