+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["dag", "primary", "node", "crypto", "worker", "consensus", "network", "config", "subscriber"]
+members = ["dag", "primary", "node", "crypto", "worker", "consensus", "network", "config"]
 
 [profile.release]
 codegen-units = 1
