+++ consensus/Cargo.toml
@@ -12,12 +12,22 @@ crypto = { path = "../crypto" }
 config = { path = "../config" }
 primary = { path = "../primary" }
 tracing = "0.1.31"
+tokio-util = { version = "0.7.0", features = ["codec"] }
+serde = { version = "1.0.136", features = ["derive"] }
+serde_bytes = "0.11.5"
+futures = "0.3.21"
+rocksdb = "0.18.0"
+bytes = "1.1.0"
+async-trait = "0.1.52"
+bincode = "1.3.3"
+store = { git = "https://github.com/mystenlabs/mysten-infra.git", package = "typed-store", rev = "808de09203d147b43d59114b8afd9e51cbcf5778"}
 
 [dev-dependencies]
 bincode = "1.3.3"
 criterion = "0.3.5"
 ed25519-dalek = { version = "1.0.1", features = ["batch", "serde"] }
 pprof = { version = "0.6.2", features = ["criterion", "flamegraph"] }
+temp_testdir = "0.2"
 
 [features]
 default = ["rand"]
