--- subscriber/Cargo.toml
@@ -1,21 +1,0 @@
-[package]
-name = "subscriber"
-version = "0.1.0"
-edition = "2021"
-license = "Apache-2.0"
-authors = ["Mysten Labs <build@mystenlabs.com>"]
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio = { version = "1.17.0", features = ["sync", "rt", "macros"] }
-serde = { version = "1.0.136", features = ["derive"] }
-serde_bytes = "0.11.5"
-log = "0.4.14"
-futures = "0.3.21"
-rocksdb = "0.18.0"
-store = { git = "https://github.com/mystenlabs/mysten-infra.git", package = "typed-store", rev = "808de09203d147b43d59114b8afd9e51cbcf5778"}
-
-crypto = { path = "../crypto" }
-primary = { path = "../primary" }
